Refactor: Factor out the reporting of test suit results to a Reporter object
